spragubr_project/requirements.txt
Flask==1.1.1
google-cloud-datastore==1.10.0  
passlib==1.7.2
pyjwt==1.7.1